kernel: Fix k_*_sys_clock_always_on macro
Commit 14e963afec0f [formerly 2b8cf4c98e44] ("include: kernel: Fix documentation for
TICKLESS_KERNEL API's")' defined a macro to fix documentation when
TKCKLESS_KERNEL is not available but this macro does not return the
same the functions returns, so its use may result in compilation
error.

Another point to consider is that if one is using this function
without it be enabled is better to return a proper error like ENOTSUP
explicitly saying that this is not supported.

Signed-off-by: Flavio Ceolin <flavio.ceolin@intel.com>

Former-commit-id: f23a8cdd2d679401f87b133be756ecf127135454