boards: add usb to stm32f4_disco
This patch adds USB capabilities to the stm32f4 discovery board.
As the board has a MicroUSB port that I would like to use as stdout I
just added USB. The needed entry within pinmux.c is already there but
USB was missing in the dts file and documentation.

I tested it with the upcomming commit.

Signed-off-by: Alexander Polleti <metapsycholo@gmail.com>

Former-commit-id: dc60f86d89af48ba321e59511ccda1a686ae08ea