misc: mempool: Fix bit mask value
SYS_MEM_POOL_KERNEL and SYS_MEM_POOL_USER are used by k_mem_pool flags
and should be defined as a bit mask.

It was defined as zero failing in all places checking for this
bit. e.g pool_irq_lock and pool_irq_unlock were not working correctly
because the check for this flag was always returning false.

Signed-off-by: Flavio Ceolin <flavio.ceolin@intel.com>

Former-commit-id: 19b7eaaa65f6354a831f02372a9e79eab0198bc0