doc: give nicer error if env not set up
If one forgets to source the zephyr-env.sh script before starting a doc
build (using the top-level Makefile), a long screenful of error messages
go by and you can miss what the actual problem is.

Check if ZEPHYR_BASE is not set and give a nice short error message.

(Updated to a more generic error message in anticipation of #9801
deprecating use of zephyr-env.sh)

Signed-off-by: David B. Kinder <david.b.kinder@intel.com>

Former-commit-id: 41bc9cf01261882dba3151a615080457b9aa0373