CMakeLists.txt: Set application shared memory dependency
The scripts for app_smem need to be ready before the script for
privilaged stack is run. This is needed because there is a small
possibility that the stacks get aligned (due to app_smem)
after privilaged stack script is run. This causes the gperf
to create an incorrect hash.

Signed-off-by: Adithya Baglody <adithya.nagaraj.baglody@intel.com>

Former-commit-id: d4b5ab6478f15f175f1aca766d460b98088142da